rename stepper to automation
