+++ .env
@@ -1,5 +1,5 @@
 # This file will be run when entering the Pipenv environment
 # Environment variables here will be available during development and Travis builds
 
-DOCKER_REPO=brewblox/brewblox-stepper
-GITHUB_REPO=BrewBlox/brewblox-stepper
+DOCKER_REPO=brewblox/brewblox-automation
+GITHUB_REPO=BrewBlox/brewblox-automation
