+++ tox.ini
@@ -1,7 +1,7 @@
 [pytest]
 flake8-max-line-length = 120
 addopts =
-    --cov=brewblox_stepper
+    --cov=brewblox_automation
     --cov-branch
     --cov-report=term-missing:skip-covered
     --cov-fail-under=100
